commit con bugs, pero con la bd actualizada
