+++ placita-app/src/index.css
@@ -10,13 +10,7 @@ body {
 }
 
 button, input, select {
-    font-family: 'Inter' !important;
-    -webkit-user-select: none;
-    /* Safari */
-    -ms-user-select: none;
-    /* IE 10 and IE 11 */
-    user-select: none;
-    /* Standard syntax */
+    font-family: 'Inter' !important;   
     
 }
